fix: split oversized files to meet 1000-line limit
- Extract template and create-event-from-template tests from
  db/tests.rs (1160 lines) into new db/template_tests.rs
- Move youtube_oauth_tests module from router.rs (1009 lines) into
  router_tests.rs where OBS tests already live
- Combine same-path routes in router.rs to reduce line count
- All three originally oversized files now under 1000 lines
