+++ crates/rs-api/src/router_tests.rs
@@ -333,3 +333,122 @@ mod obs_tests {
         assert_eq!(resp.status(), StatusCode::SERVICE_UNAVAILABLE);
     }
 }
+
+#[cfg(test)]
+mod youtube_oauth_tests {
+    use crate::router::build_router;
+    use crate::state::AppState;
+    use axum::body::Body;
+    use axum::http::{Request, StatusCode};
+    use rs_core::config::Config;
+    use rs_core::db;
+    use rs_core::models::WsEvent;
+    use tokio::sync::broadcast;
+    use tower::ServiceExt;
+
+    fn yt_config() -> Config {
+        let mut c = Config::for_testing();
+        c.youtube.client_id = "yt-cid-for-test".into();
+        c.youtube.client_secret = "yt-cs-for-test".into();
+        c
+    }
+
+    async fn test_state_with_yt() -> AppState {
+        let pool = db::create_memory_pool().await.unwrap();
+        db::run_migrations(&pool).await.unwrap();
+        let (ws_tx, _) = broadcast::channel::<WsEvent>(16);
+        AppState::new(pool, yt_config(), ws_tx)
+    }
+
+    async fn test_state_no_yt() -> AppState {
+        let pool = db::create_memory_pool().await.unwrap();
+        db::run_migrations(&pool).await.unwrap();
+        let (ws_tx, _) = broadcast::channel::<WsEvent>(16);
+        AppState::new(pool, Config::for_testing(), ws_tx)
+    }
+
+    #[tokio::test]
+    async fn oauth_start_returns_url() {
+        let state = test_state_with_yt().await;
+        let app = build_router(state);
+
+        let resp = app
+            .oneshot(
+                Request::builder()
+                    .uri("/api/v1/youtube/oauth/start")
+                    .body(Body::empty())
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+
+        assert_eq!(resp.status(), StatusCode::OK);
+        let body = axum::body::to_bytes(resp.into_body(), 1024 * 1024)
+            .await
+            .unwrap();
+        let val: serde_json::Value = serde_json::from_slice(&body).unwrap();
+        let url = val["url"].as_str().unwrap();
+        assert!(url.contains("yt-cid-for-test"));
+        assert!(url.contains("response_type=code"));
+        assert!(url.contains("access_type=offline"));
+    }
+
+    #[tokio::test]
+    async fn oauth_start_no_creds_returns_400() {
+        let state = test_state_no_yt().await;
+        let app = build_router(state);
+
+        let resp = app
+            .oneshot(
+                Request::builder()
+                    .uri("/api/v1/youtube/oauth/start")
+                    .body(Body::empty())
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+
+        assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
+    }
+
+    #[tokio::test]
+    async fn oauth_callback_no_code_returns_400() {
+        let state = test_state_with_yt().await;
+        let app = build_router(state);
+
+        let resp = app
+            .oneshot(
+                Request::builder()
+                    .uri("/api/v1/youtube/oauth/callback")
+                    .body(Body::empty())
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+
+        assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
+    }
+
+    #[tokio::test]
+    async fn oauth_callback_error_param_returns_html() {
+        let state = test_state_with_yt().await;
+        let app = build_router(state);
+
+        let resp = app
+            .oneshot(
+                Request::builder()
+                    .uri("/api/v1/youtube/oauth/callback?error=access_denied")
+                    .body(Body::empty())
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+
+        assert_eq!(resp.status(), StatusCode::OK);
+        let body = axum::body::to_bytes(resp.into_body(), 1024 * 1024)
+            .await
+            .unwrap();
+        let text = String::from_utf8(body.to_vec()).unwrap();
+        assert!(text.contains("Authorization Failed"));
+    }
+}
